requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
banal==1.0.6
beautifulsoup4==4.14.3
bleach==6.3.0
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.13
dataset==1.6.2
datasets==4.4.2
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
executing==2.2.1
fastjsonschema==2.21.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
hu_core_news_lg @ https://huggingface.co/huspacy/hu_core_news_lg/resolve/main/hu_core_news_lg-any-py3-none-any.whl#sha256=e8cd6a5478419a8cdf3af1c3e3447565a49b8c60228acf76fdbd4c5ae3611cf1
huggingface-hub==0.36.0
huspacy==0.12.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==3.13.0
kiwisolver==1.4.9
lark==1.3.1
libclang==18.1.1
lxml==6.0.2
lxml_html_clean==0.4.3
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.2
notebook==7.5.1
notebook_shim==0.2.4
numpy==2.4.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
preshed==3.0.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.48.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.54
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
terminado==0.18.1
tf_keras==2.20.1
thinc==8.3.10
threadpoolctl==3.6.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.2
wandb==0.23.1
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
